Replace RXTX-based serial library by one based on JSSC
Development on this was done in https://github.com/gohai/processing-serial-jssc. See also https://github.com/gohai/java-simple-serial-connector for changes to JSSC (not yet upstreamed).
