+++ java/libraries/serialold/library/.cvsignore
@@ -1,0 +1,1 @@
+serial.jar
