+++ java/libraries/serialold/library/.gitignore
@@ -1,0 +1,1 @@
+serialold.jar
